moving functions around to resolve circular dependency
