+++ grits/__init__.py
@@ -1,5 +1,5 @@
-from . import utils
 from .__version__ import __version__
 from .cg_compound import CG_Compound
+from . import utils
 
 __all__ = ["__version__", "CG_Compound", "utils"]
